Moved backend server to correct directory
