+++ backend/controllers/eventController.js
@@ -1,0 +1,9 @@
+const events = require('../model/TestGetEvents.js');
+
+const getEvents = (req, res) => {
+    res.json(events);
+};
+
+module.exports = {
+    getEvents,
+};
