+++ backend/model/TestGetEvents.js
@@ -1,0 +1,52 @@
+const events = [
+    {
+        "id": "1",
+        "title": "Epic Event",
+        "date": "22-12-2024",
+        "startTime": "12:30",
+        "endTime": "14:45",
+        "description": "This event will be very epic."
+    },
+    {
+        "id": "2",
+        "title": "Christmas Eve 2024 Employee Party",
+        "date": "24-12-2024",
+        "startTime": "09:30",
+        "endTime": "11:45",
+        "description": "Merry christmas! Yet another year we celebrate christmas!"
+    },
+    {
+        "id": "3",
+        "title": "Microsoft goes bankrupt celebration",
+        "date": "32-02-2069",
+        "startTime": "08:30",
+        "endTime": "23:59",
+        "description": "Destroyed by the AI of TGCD122h"
+    },
+    {
+        "id": "4",
+        "title": "AI takes over the world",
+        "date": "09-07-2077",
+        "startTime": "03:30",
+        "endTime": "07:00",
+        "description": "AI created by TGCD122h executes the final human."
+    },
+    {
+        "id": "5",
+        "title": "tiny",
+        "date": "11-11-1111",
+        "startTime": "0:00",
+        "endTime": "23:59",
+        "description": "smol event"
+    },
+    {
+        "id": "6",
+        "title": "VERY LARGE AND OBNOXIOUS EVENT",
+        "date": "12-34-56789101112",
+        "startTime": "1234:5678",
+        "endTime": "9123:18929894",
+        "description": "This event is so big it will cover a large area of any screen that tries to render this event card hehehehehehehehheheheheheheheheheheheheh"
+    },
+]
+
+module.exports = events;
