+++ backend/routes/eventRoutes.js
@@ -1,0 +1,6 @@
+const { getEvents } = require('../controllers/eventController');
+const router = express.Router();
+
+router.get('/', getEvents);
+
+module.exports = router;
