--- event-app/backend/process.env
@@ -1,3 +1,0 @@
-MONGODB_URI=mongodb://localhost:27017/mydatabase
-PORT=5000
-SECRET_KEY=mysecretkey123
